ARM: MMU64: map text segment ro and data segments execute never
With this all segments in the DRAM except the text segment are mapped
execute-never so that only the barebox code can actually be executed.
Also map the readonly data segment readonly so that it can't be
modified.

The mapping is only implemented in barebox proper. The PBL still maps
the whole DRAM rwx.

Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250618-mmu-xn-ro-v3-6-2a9825725934@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>