partitions: Start partitions at 8MiB offset
There are many SoCs that store the bootloader on the raw eMMC/SD device
outside of partitions. Examples are all i.MX SoCs and Rockchip SoCs.
Until now the first partition starts at offset 1MiB and with this we
risk conflicting with the bootloader.

It's cumbersome to sort out all SoCs that have this problem. With
nowadays eMMC/SD card sizes it doesn't really matter anymore, so change
the 1MiB offset to 8MiB which is still only a small fraction of the
card size, but leaves enough space for putting the bootloader into.

Note the change in the GPT partition handling introduced with this
change. The layout previously was:

LBA0: MBR
LBA1: GPT header
LBA2-33: GPT partition entries

which now becomes:

LBA0: MBR
LBA1: GPT header
LBA16352-16383: GPT partition entries

This means the GPT partition entries moved from LBA2 to just below the
8MiB boundary and LBA2-16351 are now free for the bootloader. This helps
with with older SoCs on which the bootloader overlaps with the normal
location of the GPT partition entries, like i.MX7 and earlier.

For nowadays disk sizes a 8MiB offset is negligible. However, for very
size constraint environments the first usable LBA is configurable with
the globalvar global.partitions.first_usable_lba which defaults to 16384
aka 8MiB.

Link: https://lore.barebox.org/20250602-createnv-v1-2-c3239ff875d5@pengutronix.de
Link: https://lore.barebox.org/20250620-createnv-v2-2-7fd3cc286b9a@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>