cdev: fix cdev_open_by_name() misuse
cdev_open_by_name() opens a cdev by its name or path. The open by path
is implemented with a simple skip the "/dev/" part of the string and use
the remaining part as name.
The effect is that several commands in barebox work with "/dev/mmc0" and
"mmc0" as argument, but not with "dev/mmc0" or even "../dev/mmc0".

Create a new cdev_open_by_path_name() function which uses
canonicalize_path() to resolve the full path and skip the "/dev/" part
afterwards.

Convert the places where a path or a name is expected to use the new
function. With this we can drop the "/dev/" skipping from
cdev_open_by_name() so that this function does what the name suggests.

Link: https://lore.barebox.org/20250602-createnv-v1-3-c3239ff875d5@pengutronix.de
Link: https://lore.barebox.org/20250620-createnv-v2-3-7fd3cc286b9a@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>