commands: create createnv command
We want to move away from describing the barebox environment explicitly
in the device tree and instead motivate usage of GPT partitions for the
envrionment. This patch creates a createnv command to facilitate this.
It creates an environment partition on the specified device and if
necessary also a GPT partition table. In the simplest case a "createnv"
without arguments will create a partition on the device barebox itself
booted from. Both the device and the size of the partition can be
specified on the command line.

We can't create environment partitions on a device containing a MBR
partition table, so a MBR partitioned device won't be touched.

As an additional safety net the command will ask the user if altering
the partition table is really desired. This behaviour can be overwritten
by specifying -f on the command line.

Link: https://lore.barebox.org/20250602-createnv-v1-4-c3239ff875d5@pengutronix.de
Link: https://lore.barebox.org/20250620-createnv-v2-4-7fd3cc286b9a@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>