mci: add option to detect non-removable cards during startup
We can optionally detect all MMC/SD cards during startup which might
cost some boot time, so this is best disabled. This adds another option
to detect only cards that are marked non-removable in the device tree
during startup. The rationale is that we will likely boot from that
device anyway, so detecting it during startup does not add to the boot
time.

We are aiming to find a good place for the barebox environment without
having to explicitly add it to the device tree. Putting the environment
into a regular partition is a good start, but the device that has the
environment needs to be available when we are searching for it. Enabling
this new option is one way to it.

Link: https://lore.barebox.org/20250602-createnv-v1-5-c3239ff875d5@pengutronix.de
Link: https://lore.barebox.org/20250620-createnv-v2-5-7fd3cc286b9a@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>