fs: fix file creation by not pointing file dentry at parent dir
A last-minute regression fix to allow unmounting file systems in some
situations unfortunately introduced a regression of its own:
Creating new files with O_CREAT associated a file descriptor's struct
file with the dentry of the parent directory instead of the dentry
of the newly created file.

Opening already existing files worked fine though, leading to e.g.
problems when saving a fresh environment to a directory:

  barebox@TI AM335x BeagleBone black:/ rm /mnt/mmc0.0/barebox.env
  barebox@TI AM335x BeagleBone black:/ saveenv
  saving environment to /mnt/mmc0.0/barebox.env
  could not open /mnt/mmc0.0/barebox.env: Invalid argument
  saveenv: Invalid argument
  barebox@TI AM335x BeagleBone black:/ saveenv
  saving environment to /mnt/mmc0.0/barebox.env

Opening existing files did not regress, because filename_lookup stores
the dentry of the actual file, unlike filename_create, which stores the
dentry of the parent directory.

Fixes: a3ef153fdf26 ("fs: unreference path on file close")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250624152148.2230298-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>