ARM: vexpress: increase barebox partition size
Commit 3630fa400be5 ("ARM: remove vexpress_defconfig") made
multi_v7_defconfig the go-to defconfig for vexpress.

The multi defconfig enables a lot of drivers and functionality though
increasing the image size from 300K to 1.2M, which causes barebox to
extend beyond the previously configured 512K barebox partition.

Fix this by giving each of barebox and its environment 2M of space,
which should be enough for many years to come.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250624183930.2733353-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>