fs: drop f_dentry
Defining f_dentry f_path.dentry is confusing as it suggests that a
struct member f_dentry is modified that doesn't exist and furthermore
it suggests that the f_path member is not modified. Drop the define
and replace its usage with f_path.dentry directly. While at it remove
the unused define f_vfsmnt.

Link: https://lore.barebox.org/20250625083544.2660810-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>