ARM: mach-imx: tzasc: add region configure helpers
At the moment the TZC380 driver is very limited and focused on minimal
platform setup unlike the TZC400 driver.

This commit adds helper functions to setup any number of TZASC regions
of any size which is required by later commits to setup an early
non-secure TZASC region1.

The code is based on the TZC400 barebox driver and the TZC380 OP-TEE
driver.

Acked-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20250630-arm-optee-early-helper-v2-7-c8cce3ae42b0@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>