ARM: i.MX: add imx6_can_access_tzasc()
On ARMv7 there is no direct way to detect if we are in the secure or non
secure world. Add a imx6_can_access_tzasc() for this purpose. When
accessing the TZASC triggers a data abort then we are in the non secure
world, because OP-TEE configures the CSU to forbid accesses to the TZASC
from the normal world. This function can be used later to detect if we
have to load OP-TEE or not.

We'll need CONFIG_ARM_EXCEPTIONS_PBL for this functionality, so select
this symbol in Kconfig.

Link: https://lore.barebox.org/20250630-arm-optee-early-helper-v2-10-c8cce3ae42b0@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>