ARM: i.MX: Webasto ccbv2: fix barebox chainloading with OP-TEE enabled
When barebox starts we have to guess if we have to start OP-TEE or not.
The current detection works by checking if the first stage passed us a
device tree pointer. This is not robust and might have security issues
[1], so replace that with the check with imx6_can_access_tzasc(). If we
can access the TZASC then we are the first stage and configure it and
start OP-TEE, otherwise assume that we are chainloaded and continue
without starting OP-TEE.

Chainloading barebox with OP-TEE enabled contained several bugs, so it
never actually worked. This patch fixes them.

Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250630-arm-optee-early-helper-v2-13-c8cce3ae42b0@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>