menu: schedule pollers/workqueues while waiting for menu item selection
Pollers are already run during the menu, because getchar() will poll the
console driver every 100us and 100us happens to be the time at which
is_timeout starts scheduling. Workqueus can only be run outside command
context however, so running menutree (e.g., because shell was exited)
will not process any fastboot packets in the workqueue.

Fix this by releasing the slice for the duration of readkey. We do a
similar thing in the readline command as well.
Unlike the readline command, menutree can execute arbitrary scripts as
a result of the user input, so we restrict the slice only to the console
and rescheduling parts.

Reported-by: Bastian Krause <bst@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Tested-by: Bastian Krause <bst@pengutronix.de>
Link: https://lore.barebox.org/20250701085251.148111-3-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>