ARM: ARM32: Fix FIT image booting
Commit 8f71bf6146acc3 ("bootm: create generic FIT image handler")
introduced a generic FIT image handler. With this ARM32 FIT images can
no longer be booted. The ARM32 FIT image handler used to call
do_bootm_linux(), but with the generic FIT image handler we select the
boot handler based on the filetype of the Kernel image which is a zImage
and consequently we call do_bootz_linux(). do_bootz_linux() operates on
data->os_file which points to the original FIT image, so booting fails
with:

invalid magic 0x28341b01

As a workaround check if we are booting a FIT image in do_bootz_linux()
and call do_bootm_linux() in that case.

Fixes: 8f71bf6146ac ("bootm: create generic FIT image handler")
Link: https://lore.barebox.org/20250701101641.549515-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>