console: fix loose coupling between getchar and is_timeout delays
We always want to run pollers while waiting in getchar().
This is already the case, because the delay chosen in getchar is the
minimum scheduling delay used in is_timeout.

To make the relation more obvious, define a constant for the minimum
scheduling delay and use it at both places.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250701131802.465283-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>