hab: writing the same srk hash is not an error
Before this commit writing the same srk hash twice failed with an error.
This is inconsistent with other invocations of the hab tool e.g.: `hab
-l -p` doesn't fail if the lockdown bit is already set.

So let's align the behavior by only throwing an error if we attempt to
write a srk hash that is different from the one that is already burned.

Signed-off-by: Stefan Kerkmann <s.kerkmann@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250709-fix-hab-equal-hash-writes-v1-1-6a59bb4b5ff1@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>