ARM: zynq: drop the board-specific environment
The Linux device tree hardcodes per DT alias (since 2015) the
stdout-path to uart1, not uart0 and thus the Linux TTY device is called
ttyPS0 and not ttyPS1.

With previous changes to the serial driver to fix up the console device,
hardcoding this option is no longer needed, so drop it along with its
environment overlay that's now empty.

Reported-by: Gwenhael Goavec-Merou <gwenhael.goavec-merou@trabucayre.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Tested-by: Gwenhael Goavec-Merou <gwenhael.goavec-merou@trabucayre.com>
Link: https://lore.barebox.org/20250714071249.2428110-2-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>