drivers: fix buggy comparisons of casted device_get_match_data to zero
device_get_match_data returns NULL when no match data is available,
which a cast to uintptr_t turns into a zero.

Most drivers either put valid pointers into the match data or they
reserve a literal value of zero to be "unknown".

Three drivers don't do that though and use 0 as valid value for the
match data and thus the check against zero breaks probing drivers with
the corresponding match data.

As device_get_match_data() would only fail anyway when manually
allocating platform devices that match by name or when adding new matches
without match data, it's perfectly fine to just skip the check and
assume the match data to be zero in that case to fix the breakage.

Fixes: 20d87123a638 ("treewide: replace dev_get_drvdata with device_get_match_data")
Reported-by: Renaud Barbier <Renaud.Barbier@ametek.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250715201908.2515578-1-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>