fastboot: populate is-logical:$partition variable
Uwe reports on IRC that fastboot tries to read some default variables
that barebox doesn't define, which triggers host-side error output:

  ERROR: fastboot: no such variable: is-logical:kernel

Flashing didn't seem impacted though.

Referring to the documentation[1], is-logical is related to a number of
commands that create/delete/resize logical partitions.

We have support for modifying partition tables in barebox, but we don't expose
it over Fastboot. Until we do, let's report all partitions as physical
partitions, which is the default value.

This issue has always existed, but only became apparent once we started
returning an error for non-existent variables.

[1]: https://android.googlesource.com/platform/system/core/+show/refs/heads/main/fastboot/README.md#237

Fixes: dd377c937f8b ("common: fastboot: send FAIL if variable does not exist")
Fixes: 3d5080aae90d ("USB: gadget: Add Android fastboot support")
Reported-by: Uwe Kleine-König <uwe@kleine-koenig.org>
Cc: Andrei Lalaev <andrey.lalaev@gmail.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250722140534.3349443-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>