commands: mmc_extcsd: fix calculation of RPMB size
EXT_CSD_RPMB_SIZE_MULT contains the number of 128 KB blocks in the RPMB.
While the calculated RMPB is labeled as KB, it actually contains the
size in bytes.

Fix the calculation to report the RPMB size in KB.

An alternative would have been to change the label to bytes, but since
the RPMB size is at least in an order of magnitude of KB, changing the
label is not intuitive.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20250729135924.122052-1-m.tretter@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>