video: remap framebuffer as writecombine if possible
We do not do cache maintenance for framebuffers and expect that memory
is mapped suitably, either writecombine or uncached (coherent).

The simplefb-client was first added to barebox for a cache-coherent
platform and did not observe this requirement.

To enable its use on ARM as well without graphic artifacts, let's remap
the range as writecombine, which will fall back to uncached if
unsupported.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250730123126.1062875-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>