cdev: fix mtd partition handling
devfs-core.c implements the partition handling only for devices that are
not MTD devices. MTD devices have their own partition handling in
drivers/mtd/partition.c, see the special case in __devfs_add_partition()
for MTD devices. This means for MTD devices we have to pass the original
cdev to the ops, not the master cdev.

Fixes: 6d2b6f399c07 ("cdev: pass master cdev to cdev ops")
Tested-by: Stefano Manni <stefano.manni@gmail.com>
Link: https://lore.barebox.org/20250805072537.3484221-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>