ARM: mmu: setup trap pages before remapping R/O
The order matters on ARM32, because arm_fixup_vectors() actually rewrite
the vector table, which is in the text area.

On ARM64, the order doesn't matter. As we are going to make the memory
bank remapping code common between both ARM32 and ARM64, move
setup_trap_pages, so the code between ARM32 and ARM64 becomes identical.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250805174541.2606267-7-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>