ARM: mmu64: rename __find_pte to find_pte
find_pte is useful for debugging, but was unreferenced at first until we
added flush_cacheable_pages as user of __find_pte.

__find_pte is easy to use, so just drop find_pte and rename __find_pte
to find_pte. This is in preparation for mmu32 to provide a function
with the same semantics.

While at it, change the level to be unsigned as there are no negative
levels.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250806123714.2092620-17-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>