driver: introduce common struct bobject
We do not have Linux kobject with attributes in barebox, but we do
have something roughly equivalent: struct device and device parameters,
which enables us to do things like eth0.mode=dhcp on the shell.

The parameter API is fairly expressive and would be very useful for
transferring data between commands. In preparation for doing that,
introduce struct bobject, which will allow device-less parameters in
the future. We keep it in a union with the name, so we do not have to
patch all instances of dev->name across the tree.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250811122824.1667791-3-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>