param: operate on bobjects instead of full devices
Previous commit associated the parameters with the bobject inside the
device. Double down and provide bobject API for creation and use of
parameters and use them to implement the device parameter API.

As each device starts with a bobject, let's make use of GCC's
transparent union feature (typedeffed as bobject_t) to have the compiler
automatically convert pointers.
This way the dev_ param functions become simple aliases to the bobject_
API, but the main upside is that we do not need to adjust callbacks
across the tree as they will just continue to work without type errors.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250811122824.1667791-8-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>