efi: fix probing of EFI devices with parents
The refactor in the referenced commit sends in the match function as
the context data - instead of the handle we are looking for.

This meant that efi_find_device() always returned NULL, leading to all
EFI devices with parent handles not being probed.

Fix this by sending in the handle instead.

Fixes: 4a5a9b58b568 ("drive: bus: make use of new bus_find_device helper")
Signed-off-by: Tobias Waldekranz <tobias@waldekranz.com>
Tested-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250813214103.254178-1-tobias@waldekranz.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>