kbuild: images/Makefile: fix broken if_changed usage
if_changed also triggers when the command line changes by comparing the
command executed against what was written last time.

This necessarily means that there may only be a single if_changed called
per target, a rule which we violated.

This led to breakage on RISC-V, because every build triggered a relink
of the PBLs, but only every second build executed cmd_prelink__.

Switch over to if_changed_rule to fix this. This also means we need to
move the comments out of the define, so they aren't printed.

Fixes: 912696b571bb ("images: don't compute size if input file didn't change")
Fixes: 4c31478919f5 ("images: don't prelink if input file didn't change")
Fixes: c2e320f6b243 ("images: don't check for missing FW if file didn't change")
Fixes: 8ddc2d799b4a ("images: don't compare filesizes if file didn't change")
Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250814202254.344010-1-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>