RISC-V: virt: riscvemu: set better hostname/model
The hostname and model currently come from the device tree:
"riscvemu-bar_dev" and "ucbbar,riscvemu-bare", respectively.

The former is now warned about, because it contains an underscore and
the latter is not that descriptive.

As the device tree is supplied by the emulator, add an override to board
code. We differntiate between riscv32 and riscv64 in the hostname string
as that information can be relevant for choosing the correct kernel.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250814202805.354827-2-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>